requirements.txt
requests
anthropic
openai
python-dotenv
rich
google-genai
groq
